requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118
numpy
pandas
matplotlib
transformers 
sentencepiece 
datasets[audio]
pytube
speechbrain
moviepy
pyannote.audio
ipywidgets
whisperx
ffmpeg
whisper
TTS
gradio
openai
tiktoken
chromadb
langchain_experimental
langchain[docarray]
beautifulsoup4==4.12.3
langchain==0.1.12
langchain_community==0.0.28
langchain_core==0.1.32
PyPDF2==3.0.1
Requests==2.31.0
python-dotenv
